Long -> Integer

+++ src/org/labkey/test/AssayAPITest.java
@@ -161,7 +161,7 @@ public class AssayAPITest extends BaseWebDriverTest
                 .setEditableRuns(true);
         SaveProtocolCommand saveProtocolCommand = new SaveProtocolCommand(newAssayProtocol);
         ProtocolResponse saveProtocolResponse = saveProtocolCommand.execute(connection, getCurrentContainerPath());
-        Long protocolId = saveProtocolResponse.getProtocol().getProtocolId();
+        Integer protocolId = saveProtocolResponse.getProtocol().getProtocolId();
 
         assertEquals(assayDescription, saveProtocolResponse.getProtocol().getDescription());
         assertTrue(saveProtocolResponse.getProtocol().getQcEnabled());
@@ -187,7 +187,7 @@ public class AssayAPITest extends BaseWebDriverTest
         APIAssayHelper assayHelper = new APIAssayHelper(this);
         int assayId = assayHelper.getIdFromAssayName(assayName, getProjectName(), false);
         if (assayId == 0)
-            assayId = assayHelper.createAssayDesignWithDefaults(getProjectName(), "General", assayName).getProtocolId().intValue();
+            assayId = assayHelper.createAssayDesignWithDefaults(getProjectName(), "General", assayName).getProtocolId();
 
         // First, simulate file already being uploaded to the server by copying to the pipeline root
         List<String> lines1 = Arrays.asList(
